refactor(query): Refactor QueryInputBox into a dedicated component from SearchTabPanel.
