+++ src/components/CentralContainer/Sidebar/SidebarTabs/SearchTabPanel/index.tsx
@@ -1,4 +1,4 @@
-import React, {
+import {
     useContext,
     useState,
 } from "react";
@@ -6,9 +6,6 @@ import React, {
 import {
     AccordionGroup,
     Box,
-    LinearProgress,
-    Stack,
-    Textarea,
 } from "@mui/joy";
 
 import UnfoldLessIcon from "@mui/icons-material/UnfoldLess";
@@ -16,19 +13,13 @@ import UnfoldMoreIcon from "@mui/icons-material/UnfoldMore";
 
 import {StateContext} from "../../../../../contexts/StateContextProvider";
 import {
-    QUERY_PROGRESS_VALUE_MAX,
-    QueryArgs,
-} from "../../../../../typings/query";
-import {UI_ELEMENT} from "../../../../../typings/states";
-import {
     TAB_DISPLAY_NAMES,
     TAB_NAME,
 } from "../../../../../typings/tab";
-import {isDisabled} from "../../../../../utils/states";
 import CustomTabPanel from "../CustomTabPanel";
 import PanelTitleButton from "../PanelTitleButton";
+import QueryInputBox from "./QueryInputBox";
 import ResultsGroup from "./ResultsGroup";
-import ToggleIconButton from "./ToggleIconButton";
 
 import "./index.css";
 
@@ -39,41 +30,13 @@ import "./index.css";
  * @return
  */
 const SearchTabPanel = () => {
-    const {queryProgress, queryResults, startQuery, uiState} = useContext(StateContext);
+    const {queryResults} = useContext(StateContext);
+
     const [isAllExpanded, setIsAllExpanded] = useState<boolean>(true);
-    const [queryString, setQueryString] = useState<string>("");
-    const [isCaseSensitive, setIsCaseSensitive] = useState<boolean>(false);
-    const [isRegex, setIsRegex] = useState<boolean>(false);
 
     const handleCollapseAllButtonClick = () => {
         setIsAllExpanded((v) => !v);
     };
-
-    const handleQuerySubmit = (newArgs: Partial<QueryArgs>) => {
-        startQuery({
-            isCaseSensitive: isCaseSensitive,
-            isRegex: isRegex,
-            queryString: queryString,
-            ...newArgs,
-        });
-    };
-
-    const handleQueryInputChange = (ev: React.ChangeEvent<HTMLTextAreaElement>) => {
-        setQueryString(ev.target.value);
-        handleQuerySubmit({queryString: ev.target.value});
-    };
-
-    const handleCaseSensitivityButtonClick = () => {
-        handleQuerySubmit({isCaseSensitive: !isCaseSensitive});
-        setIsCaseSensitive(!isCaseSensitive);
-    };
-
-    const handleRegexButtonClick = () => {
-        handleQuerySubmit({isRegex: !isRegex});
-        setIsRegex(!isRegex);
-    };
-
-    const isQueryInputBoxDisabled = isDisabled(uiState, UI_ELEMENT.QUERY_INPUT_BOX);
 
     return (
         <CustomTabPanel
@@ -93,59 +56,7 @@ const SearchTabPanel = () => {
             }
         >
             <Box className={"search-tab-container"}>
-                <div className={"query-input-box-with-progress"}>
-                    <Textarea
-                        className={"query-input-box"}
-                        maxRows={7}
-                        placeholder={"Search"}
-                        size={"sm"}
-                        endDecorator={
-                            <Stack
-                                direction={"row"}
-                                spacing={0.25}
-                            >
-                                <ToggleIconButton
-                                    className={"query-option-button"}
-                                    disabled={isQueryInputBoxDisabled}
-                                    isChecked={isCaseSensitive}
-                                    size={"sm"}
-                                    tooltipTitle={"Match case"}
-                                    variant={"plain"}
-                                    onClick={handleCaseSensitivityButtonClick}
-                                >
-                                    Aa
-                                </ToggleIconButton>
-
-                                <ToggleIconButton
-                                    className={"query-option-button"}
-                                    disabled={isQueryInputBoxDisabled}
-                                    isChecked={isRegex}
-                                    size={"sm"}
-                                    tooltipTitle={"Use regular expression"}
-                                    variant={"plain"}
-                                    onClick={handleRegexButtonClick}
-                                >
-                                    .*
-                                </ToggleIconButton>
-                            </Stack>
-                        }
-                        slotProps={{
-                            textarea: {
-                                className: "query-input-box-textarea",
-                                disabled: isQueryInputBoxDisabled,
-                            },
-                            endDecorator: {className: "query-input-box-end-decorator"},
-                        }}
-                        onChange={handleQueryInputChange}/>
-                    <LinearProgress
-                        className={"query-input-box-linear-progress"}
-                        determinate={true}
-                        thickness={4}
-                        value={queryProgress * 100}
-                        color={QUERY_PROGRESS_VALUE_MAX === queryProgress ?
-                            "success" :
-                            "primary"}/>
-                </div>
+                <QueryInputBox/>
                 <AccordionGroup
                     className={"query-results"}
                     disableDivider={true}
